Move mart.utils.lightning.flatten_dict to mart.utils.
