+++ mart/utils/__init__.py
@@ -4,6 +4,7 @@ from .imports import _HAS_LIGHTNING
 from .monkey_patch import *
 from .optimization import *
 from .silent import *
+from .utils import *
 
 if _HAS_LIGHTNING:
     from .lightning import *
